pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aura-mcp"
version = "0.3.0"
description = "Autonomous Unified Resource Architect — an extensible MCP platform that reads Notion tasks and executes them"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["mcp", "notion", "automation", "scaffolding", "openai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "mcp>=1.0",
    "notion-client>=2.0",
    "openai>=1.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "typer>=0.9",
    "rich>=13.0",
]

[project.scripts]
aura = "aura_mcp.cli.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["aura_mcp"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
